tract3D, from vox to mm space
Tract3D used to plot images in voxel coordinates. It has now been
updated to read and apply the affine transformation matrix
(sform/qform) from the nifti images. This also means that all input to
S.surf.vol and the sorts now need to be specified in mm. Please also
note that the x and y dimensions are now swapped to match MRI (and not
matlab) conventions.