tract3D: default MNI brain
you can now specify the structural image as ‘mni’ instead of specifying
the actualy file location. It will retrieve the MNI brain image in the
FSL directory. This update also changed default threshold and color
setting for the structural (to match the MNI) and adds ‘lightangle’ to
the lighting options.